# chapter05/project/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Custom Bikes",
        "contact": {
            "name": "Developer Contact",
            "url": "https://www.example.com"
        },
        "version": 1
    },
    "host": "localhost:8000",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create new User",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success: operation Successfully"
                    },
                    "401": {
                        "description": "Refused: Unauthenticated"
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "loggin a user",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: operation Successfully"
                    },
                    "401": {
                        "description": "Refused: Unauthenticated"
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Logout a User",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: operation Successfully"
                    },
                    "401": {
                        "description": "Refused: Unauthenticated"
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid Input"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/bikes": {
            "get": {
                "tags": [
                    "Bikes"
                ],
                "summary": "List Bikes",
                "responses": {
                    "200": {
                        "description": "Success: List all Bikes",
                        "schema": {
                            "$ref": "#/definitions/Bike"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Bikes"
                ],
                "summary": "Create Bike",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Bike"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success: A Newly Created Bike",
                        "schema": {
                            "$ref": "#/definitions/Bike"
                        }
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            }
        },
        "/api/bikes/{id}": {
            "get": {
                "tags": [
                    "Bikes"
                ],
                "summary": "Get Bike by Id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Display the specified Bike by id.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Bike",
                        "schema": {
                            "$ref": "#/definitions/Bike"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            },
            "put": {
                "tags": [
                    "Bikes"
                ],
                "summary": "Update Bike",
                "description": "Update the specified Bike by Id.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bike id to update",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Bike"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Bike updated",
                        "schema": {
                            "$ref": "#/definitions/Bike"
                        }
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Bikes"
                ],
                "summary": "Delete Bike",
                "description": "Delete the specified Bike by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bike id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success: successful deleted"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Bike": {
            "required": [
                "make",
                "model",
                "year",
                "mods"
            ],
            "properties": {
                "make": {
                    "description": "Company name",
                    "type": "string",
                    "example": "Harley Davidson, Honda, Yamaha"
                },
                "model": {
                    "description": "Motorcycle model",
                    "type": "string",
                    "example": "Xl1200, Shadow ACE, V-Star"
                },
                "year": {
                    "description": "Fabrication year",
                    "type": "string",
                    "example": "2009, 2008, 2007"
                },
                "mods": {
                    "description": "Motorcycle description of modifications",
                    "type": "string",
                    "example": "New exhaust system"
                },
                "picture": {
                    "description": "Bike image url",
                    "type": "string",
                    "example": "http://www.sample.com/my.bike.jpg"
                }
            }
        },
...
        "User": {
            "required": [
                "name",
                "email",
                "password"
            ],
            "properties": {
                "name": {
                    "description": "User name",
                    "type": "string",
                    "example": "John Coner"
                },
                "email": {
                    "description": "Email Address",
                    "type": "string",
                    "example": "john.conor@terminator.com"
                },
                "password": {
                    "description": "A very secure password",
                    "type": "string",
                    "example": "tHf682k4"
                }
            }
        }
    }
}